Editor: Merge the Layout selector into the KeyPicker
